pyproject.toml
[tool.poetry]
name = "handshake-client"
version = "0.1.0"
description = "Handshake API client by Python"
authors = ["naoki-maeda"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.23.0"
python-bitcoinrpc = "^1.0"

[tool.poetry.dev-dependencies]
mypy = "^0.770"
flake8 = "^3.7.9"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
